Feature for issue #10
- Promisify Mongoose response in the controllers
- Increased readability
- Adapted tests
