+++ lib/index.js
@@ -14,6 +14,7 @@ app.use(bodyParser.urlencoded({
         extended: false
     }));
 app.use(bodyParser.json());
+mongoose.Promise = global.Promise;
 
 function initHttp() {
     return new Promise(function(resolve, reject) {
